requirements.txt
sentence-transformers
faiss-cpu
PyPDF2
numpy
huggingface-hub
